# schedule-service/docs/swagger.yaml
basePath: /raspyx/api
definitions:
  models.AddGroupRequest:
    properties:
      number:
        example: 221-352
        type: string
    type: object
...
  models.AddRoomSchedule:
    properties:
      room_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.AddScheduleRequest:
    properties:
      end_date:
        example: "2025-12-31"
        type: string
      end_time:
        example: "10:30:00"
        type: string
      group_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      is_session:
        example: false
        type: boolean
      link:
        example: https://rasp.dmami.ru/
        type: string
      location_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      rooms_uuid:
        items:
          $ref: '#/definitions/models.AddRoomSchedule'
        type: array
      start_date:
        example: "2025-09-01"
        type: string
      start_time:
        example: "09:00:00"
        type: string
      subject_type_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      subject_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
      teachers_uuid:
        items:
          $ref: '#/definitions/models.AddTeacherSchedule'
        type: array
      weekday:
        example: 1
        type: integer
    type: object
...
  models.AddTeacherSchedule:
    properties:
      teacher_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Group:
    properties:
      number:
        example: 221-352
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.GroupResponse:
    properties:
      group_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.ResponseAPI:
    properties:
      errors:
        $ref: '#/definitions/models.ErrorResponse'
      message:
        type: string
      requestId:
        type: string
      result: {}
      success:
        type: boolean
    type: object
...
  models.RoomResponse:
    properties:
      room_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.ScheduleResponse:
    properties:
      schedule_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.UpdateRoomRequest:
    properties:
      number:
        example: ав4810
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for schedules
  title: Raspyx
  version: 2.0.0
paths:
  /v2/groups:
    get:
      consumes:
      - application/json
      description: Получение списка всех учебных групп в системе
      produces:
      - application/json
      responses:
        "200":
          description: Данные об учебных группах успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.Group'
                  type: array
              type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение учебных групп
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Создание новой учебной группы
      parameters:
      - description: Данные новой учебной группы
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Учебная группа успешно создана
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.GroupResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при создании учебной группы
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Добавление учебной группы
      tags:
      - groups
...
  /v2/groups/number/{group_number}:
    get:
      consumes:
      - application/json
      description: Получение информации об учебной группе по её group_number
      parameters:
      - description: Номер учебной группы
        in: path
        name: group_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные об учебной группе успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.Group'
                  type: array
              type: object
        "404":
          description: Учебная группа не найдена
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение учебной группы по номеру
      tags:
      - groups
  /v2/groups/uuid/{group_uuid}:
    get:
      consumes:
      - application/json
      description: Получение информации об учебной группе по её group_uuid
      parameters:
      - description: Идентификатор учебной группы
        in: path
        name: group_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные об учебной группе успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.Group'
              type: object
        "404":
          description: Учебная группа не найдена
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение учебной группы по идентификатору
      tags:
      - groups
...
  /v2/rooms/{room_uuid}:
    delete:
      consumes:
      - application/json
      description: Удаление аудитории из системы по её идентификатору
      parameters:
      - description: Идентификатор аудитории
        in: path
        name: room_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Аудитория успешно удалена
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.RoomResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Аудитория не найдена
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при удалении Аудитории
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Удаление аудитории
      tags:
      - rooms
    put:
      consumes:
      - application/json
      description: Обновление информации об аудитории по её идентификатору
      parameters:
      - description: Идентификатор аудитории
        in: path
        name: room_uuid
        required: true
        type: string
      - description: Данные для обновления аудитории
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные об аудитории успешно обновлены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.RoomResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Аудитория не найдена
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при обновлении аудитории
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Обновление аудитории
      tags:
      - rooms
...
  /v2/schedule:
    post:
      consumes:
      - application/json
      description: Создание нового расписания
      parameters:
      - description: Данные нового расписания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Расписание успешно создано
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.ScheduleResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при создании расписания
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Добавление расписания
      tags:
      - schedule
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"